app/requirements.txt
psycopg2-binary==2.8.3
Django==2.1.11
djangorestframework==3.9.1
django-redis==4.11.0 
python-dateutil==2.8.1 
